Remove the YamlReaders folder in the Reader project and change the namespace
