+++ src/Microsoft.OpenApi.Readers/OpenApiDiagnostic.cs
@@ -3,12 +3,13 @@
 //  Licensed under the MIT License (MIT). See LICENSE in the repo root for license information.
 // ------------------------------------------------------------
 
-using System;
 using System.Collections.Generic;
+using Microsoft.OpenApi.Readers.Interface;
 
-namespace Microsoft.OpenApi.Readers.YamlReaders.ParseNodes
+namespace Microsoft.OpenApi.Readers
 {
-    internal class PatternFieldMap<T> : Dictionary<Func<string, bool>, Action<T, string, ParseNode>>
+    public class OpenApiDiagnostic : IDiagnostic
     {
+        public IList<OpenApiError> Errors { get; set; } = new List<OpenApiError>();
     }
 }
