+++ src/Microsoft.OpenApi.Readers/ParseNodes/FixedFieldMap.cs
@@ -6,7 +6,7 @@
 using System;
 using System.Collections.Generic;
 
-namespace Microsoft.OpenApi.Readers.YamlReaders.ParseNodes
+namespace Microsoft.OpenApi.Readers.ParseNodes
 {
     internal class FixedFieldMap<T> : Dictionary<string, Action<T, ParseNode>>
     {
