+++ src/Microsoft.OpenApi.Readers/ParseNodes/ListNode.cs
@@ -10,7 +10,7 @@ using System.Linq;
 using SharpYaml.Serialization;
 using Microsoft.OpenApi.Any;
 
-namespace Microsoft.OpenApi.Readers.YamlReaders.ParseNodes
+namespace Microsoft.OpenApi.Readers.ParseNodes
 {
     internal class ListNode : ParseNode, IEnumerable<ParseNode>
     {
