+++ src/Microsoft.OpenApi.Readers/ParseNodes/MapNode.cs
@@ -12,7 +12,7 @@ using SharpYaml.Serialization;
 using Microsoft.OpenApi.Any;
 using Microsoft.OpenApi.Interfaces;
 
-namespace Microsoft.OpenApi.Readers.YamlReaders.ParseNodes
+namespace Microsoft.OpenApi.Readers.ParseNodes
 {
     /// <summary>
     /// Abstraction of a Map to isolate semantic parsing from details of
