+++ src/Microsoft.OpenApi.Readers/ParseNodes/PropertyNode.cs
@@ -9,7 +9,7 @@ using System.Linq;
 using SharpYaml.Serialization;
 using Microsoft.OpenApi.Any;
 
-namespace Microsoft.OpenApi.Readers.YamlReaders.ParseNodes
+namespace Microsoft.OpenApi.Readers.ParseNodes
 {
     internal class PropertyNode : ParseNode
     {
