+++ src/Microsoft.OpenApi.Readers/ParsingContext.cs
@@ -8,7 +8,7 @@ using System.Linq;
 using Microsoft.OpenApi.Interfaces;
 using Microsoft.OpenApi.Models;
 
-namespace Microsoft.OpenApi.Readers.YamlReaders
+namespace Microsoft.OpenApi.Readers
 {
     public class ParsingContext
     {
