+++ src/Microsoft.OpenApi.Readers/ReferenceService.cs
@@ -1,9 +1,8 @@
 ﻿using System;
-using Microsoft.OpenApi;
 using Microsoft.OpenApi.Interfaces;
 using Microsoft.OpenApi.Models;
 
-namespace Microsoft.OpenApi.Readers.YamlReaders
+namespace Microsoft.OpenApi.Readers
 {
     internal class OpenApiReferenceService : IOpenApiReferenceService
     {
