+++ test/Microsoft.OpenApi.Readers.Tests/OpenApiExampleTests.cs
@@ -6,7 +6,6 @@
 using System;
 using System.Net.Http;
 using System.Threading.Tasks;
-using Microsoft.OpenApi.Readers.YamlReaders;
 using Xunit;
 
 namespace Microsoft.OpenApi.Readers.Tests
