+++ test/Microsoft.OpenApi.Readers.Tests/V2Tests/V2Tests.cs
@@ -3,7 +3,6 @@
 //  Licensed under the MIT License (MIT). See LICENSE in the repo root for license information.
 // ------------------------------------------------------------
 
-using Microsoft.OpenApi.Readers.YamlReaders;
 using Xunit;
 
 namespace Microsoft.OpenApi.Readers.Tests.V2Tests
